requirements.txt
# Core dependencies for RAG Research Assistant

# Web requests and API access
requests>=2.28.0
urllib3>=1.26.0

# Data processing and analysis
numpy>=1.21.0
pandas>=1.5.0

# Natural language processing
nltk>=3.8.0
spacy>=3.5.0

# Machine learning and embeddings
scikit-learn>=1.2.0
sentence-transformers>=2.2.0

# Vector database and similarity search
chromadb>=0.4.0
faiss-cpu>=1.7.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0

# XML and HTML parsing
lxml>=4.9.0
beautifulsoup4>=4.11.0
feedparser>=6.0.0

# Database - sqlite3 is built-in with Python

# Logging and monitoring
structlog>=23.1.0

# Date and time handling
python-dateutil>=2.8.0

# Text processing utilities
textdistance>=4.5.0
fuzzywuzzy>=0.18.0

# HTTP client with retry capabilities
httpx>=0.24.0

# Progress bars and CLI utilities
tqdm>=4.65.0
rich>=13.0.0

# Async support
aiohttp>=3.8.0
# asyncio is built-in with Python

# Data validation
pydantic>=2.0.0

# Caching
diskcache>=5.6.0

# System monitoring
psutil>=5

# Web interface
flask>=2.3.0
jinja2>=3.1.0
werkzeug>=2.3.0